Fix lint warnings.
